% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimation.R
\name{ppi_mean}
\alias{ppi_mean}
\title{Estimate proportion of binary outcome in full data set}
\usage{
ppi_mean(labeled_data_Y, labeled_data_predictions, unlabeled_data_predictions)
}
\arguments{
\item{labeled_data_Y}{a vector of binary outcomes (treated as ground-truth labels)}

\item{labeled_data_predictions}{a vector of predicted binary outcomes (from e.g. an LLM) for the labeled data}

\item{unlabeled_data_predictions}{a vector of predicted binary outcomes for corresponding unlabeled data}
}
\value{
a list containing the point estimate, standard error, and confidence interval. The mean of the labeled data alone is also returned.
}
\description{
This function takes a vector of observed binary outcomes (for a labeled dataset), a vector of predicted outcomes
(for the same labeled dataset), and a vector of predicted outcomes for an unlabeled dataset. It returns an estimate
of the proportion of positive binary outcomes in the full dataset (labeled and unlabeled combined). The method implemented here
follows Miao, Miao, Wu, Zhao, and Lu (2024), calculating an optimal weight for adjusting the sample mean in the
labeled data by the difference between the mean predicted labels in the labeled and unlabeled datasets.
The output also provides 95\% confidence intervals based on the asymptotic variance of the estimator.
}
