% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimation.R
\name{print.ppi_mean_estimate}
\alias{print.ppi_mean_estimate}
\title{Print method for ppi_mean_estimate class}
\usage{
\method{print}{ppi_mean_estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ppi_mean_estimate}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for ppi_mean_estimate class
}
